Sounds added; Input System; Small improvements;

+++ Technical Test Lite v2/Assets/_Project/Scripts/MapCreator.cs
@@ -1,3 +1,4 @@
+#if UNITY_EDITOR
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
@@ -5,7 +6,8 @@ using System.Text.RegularExpressions;
 using UnityEditor;
 using UnityEngine;
 
-public class MapCreator : MonoBehaviour, ISerializationCallbackReceiver
+public class MapCreator : MonoBehaviour 
+    , ISerializationCallbackReceiver
 {
     private static float tileSize = 15f;
 
@@ -25,7 +27,7 @@ public class MapCreator : MonoBehaviour, ISerializationCallbackReceiver
         maps = Resources.LoadAll<Map>("Maps").ToList();
     }
 
-    private async void CreateRandomMap(bool isNew)
+    private void CreateRandomMap(bool isNew)
     {
         if(!isNew && maps.Count == 0)
         {
@@ -48,7 +50,8 @@ public class MapCreator : MonoBehaviour, ISerializationCallbackReceiver
         tiles[0] = 0;
         for (int i = 1; i < newMapTilesCount - 1; i++)
         {
-            tiles[i] = Random.Range(0, 2);
+            int random = Random.Range(0, 50);
+            tiles[i] = random > 45 ? 1 : 0;
             if(tiles[i] == 0)
             {
                 if(first)
@@ -57,8 +60,8 @@ public class MapCreator : MonoBehaviour, ISerializationCallbackReceiver
                     first = false;
                     continue;
                 }
-                int random = Random.Range(0, 100);
-                if(random > 50)
+                random = Random.Range(0, 100);
+                if(random > 70)
                 {
                     objects[i] = 1;
                 }
@@ -148,16 +151,19 @@ public class MapCreator : MonoBehaviour, ISerializationCallbackReceiver
     // Create (or Update) the map data (on the) scriptable object
     void CreateOrUpdateScriptableObject(bool isNew)
     {
+        if(mapSpawner)
         if(isNew)
         {
             Map newMap = ScriptableObject.CreateInstance<Map>();
             string assetPath = $"Assets/Resources/Maps/Map {maps.Count}.asset";
+            newMap.UpdateData(tiles, positions, CalculatePath(), rotations, objects);
             AssetDatabase.CreateAsset(newMap, assetPath);
             level = newMap;
+            Level = $"Map {maps.Count}";
         }
-        level.UpdateData(tiles, positions, CalculatePath(), rotations, objects);
+        else maps[mapSpawner.currentLevel].UpdateData(tiles, positions, CalculatePath(), rotations, objects);
         LoadAllMaps();
-        Level = $"Map {maps.IndexOf(level)}";
+        thisMapTilesCount = maps[GetLevelNumber(Level)].map.Length;
         ShowLevelOnEditor();
     }
 
@@ -180,15 +186,15 @@ public class MapCreator : MonoBehaviour, ISerializationCallbackReceiver
                 Vector3 oldPos = path[i + counter];
                 if(rotations[i] == 0 || rotations[i] == 270)
                 {
-                    path.Insert(i + 1 + counter, oldPos - (oldPos - MoveForward(oldPos))/2);
+                    path.Insert(i + 1 + counter, oldPos - (oldPos - MoveForward(oldPos)) / 2);
                 }
                 else if(rotations[i] == 180)
                 {
-                    path.Insert(i + 1 + counter, oldPos - (oldPos - MoveUp(oldPos))/2);
+                    path.Insert(i + 1 + counter, oldPos - (oldPos - MoveUp(oldPos)) / 2);
                 }
                 else
                 {
-                    path.Insert(i + 1 + counter, oldPos - (oldPos - MoveDown(oldPos))/2);
+                    path.Insert(i + 1 + counter, oldPos - (oldPos - MoveDown(oldPos)) / 2);
                 }
                 counter++;
             }
@@ -200,13 +206,6 @@ public class MapCreator : MonoBehaviour, ISerializationCallbackReceiver
             path.Insert(i + 1 + counter, path[i + counter] + path[i + 1 + counter] - path[i + counter]);
             counter++;
         }
-        // counter = 0;
-        // listCount = path.Count - 1;
-        // for (int i = 0; i < listCount; i++)
-        // {
-        //     path.Insert(i + 1 + counter, path[i + counter] + path[i + 1 + counter] - path[i + counter]);
-        //     counter++;
-        // }
         return path.ToArray();
     }
 
@@ -214,11 +213,9 @@ public class MapCreator : MonoBehaviour, ISerializationCallbackReceiver
     {
         if(mapSpawner == null) mapSpawner = FindObjectOfType<MapSpawner>();
 
-        LoadAllMaps();
-
         if(maps != null && maps.Count > 0)
         {
-            int levelIndex = maps.IndexOf(level);
+            int levelIndex = GetLevelNumber(Level);
             mapSpawner.currentLevel = levelIndex;
             
             mapSpawner.SpawnMap();
@@ -227,16 +224,17 @@ public class MapCreator : MonoBehaviour, ISerializationCallbackReceiver
 
     private void DeleteMapData()
     {
-        int levelIndex = maps.IndexOf(level);
+        int levelIndex = GetLevelNumber(Level);
+
+        if(levelIndex < 0) return;
+
         string assetPath = $"Assets/Resources/Maps/Map {levelIndex}.asset";
         AssetDatabase.DeleteAsset(assetPath);
-        AssetDatabase.SaveAssets();
         
         for (int i = levelIndex + 1; i < maps.Count; i++)
         {
             assetPath =  $"Assets/Resources/Maps/Map {i}.asset";
             AssetDatabase.RenameAsset(assetPath, $"Map {i-1}.asset");
-            AssetDatabase.SaveAssets();
         }
         LoadAllMaps();
 
@@ -248,9 +246,11 @@ public class MapCreator : MonoBehaviour, ISerializationCallbackReceiver
             }
             level = maps[levelIndex];
             Level = $"Map {levelIndex}";
+            ShowLevelOnEditor();
         }
         else
         {
+            if(mapSpawner == null) mapSpawner = FindObjectOfType<MapSpawner>();
             Level = null;
             mapSpawner.DeleteMap();
         }
@@ -269,9 +269,13 @@ public class MapCreator : MonoBehaviour, ISerializationCallbackReceiver
         return new Vector3(currentPos.x - tileSize, currentPos.y, currentPos.z);
     }
 
-    #region Editor
+    private int GetLevelNumber(string level)
+    {
+        if(level == null) return -1;
+        return int.Parse(Regex.Match(level, @"\d+").Value);
+    }
 
-#if UNITY_EDITOR
+    #region Editor
 
         #region Popup List
 
@@ -287,7 +291,7 @@ public class MapCreator : MonoBehaviour, ISerializationCallbackReceiver
         public void OnBeforeSerialize()
         {
             LoadAllMaps();
-            PopupList = maps.OrderBy(q => int.Parse(Regex.Match(q.name, @"\d+").Value)).Select(q => q.name).ToList();
+            PopupList = maps.OrderBy(q => GetLevelNumber(q.name)).Select(q => q.name).ToList();
             TMPList = PopupList;
         }
         public void OnAfterDeserialize() {}
@@ -296,13 +300,10 @@ public class MapCreator : MonoBehaviour, ISerializationCallbackReceiver
         {
             if(maps != null && maps.Count > 0)
             {
-                Map newLevel = maps[int.Parse(Regex.Match(Level, @"\d+").Value)];
-                if(level != newLevel)
-                {
-                    level = newLevel;
-                    ShowLevelOnEditor();
-                    thisMapTilesCount = level.map.Length;
-                }
+                Map newLevel = maps[GetLevelNumber(Level)];
+                level = newLevel;
+                ShowLevelOnEditor();
+                thisMapTilesCount = level.map.Length;
             }
         }
 
@@ -341,11 +342,11 @@ public class MapCreator : MonoBehaviour, ISerializationCallbackReceiver
             }
         }
 
-#endif
 
     #endregion
 
 }
+#endif
 
 #region TileType enum
